# chatbot-ui/services/swagger-type/api-spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Huniverse",
    "description": "Hide API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/wedatalab/api"
    }
  ],
  "paths": {
...
    "/admin/manual/": {
      "get": {
        "tags": ["admin-manual"],
        "summary": "Get All",
        "description": "## 전체 메뉴얼 목록 paginate 하여 보여주기\n\n### Args:\n    page_no: int = 요청할 페이지 번호\n    page_size: int = 한 번에 조회 할 메뉴얼 수, 기본 값 50개 (10/30/50/100)\n    screen_id: Optional[str] = 화면(ID)로 검색\n    manual_name Optional[str] = 메뉴얼 이름으로 검색\n    start_dt, end_dt: Optional[float] = 메뉴얼 변동 일자 (unix time) 넣어주시면 됩니다\n\n### Raises:\n\n### Returns:\n    status: (200, 'OK')\n    data: [\n        {\n            \"id\": 1,\n            \"screen_id\": \"~~\",\n            \"manual_name\": \"00.~~pdf\",\n            \"manual_path\": \"www.~~~.kr\",\n            \"creation_dt\": 1726055378,\n            \"update_dt\": 1726055378\n        },\n        ...\n    ]",
        "operationId": "get_all_admin_manual__get",
        "parameters": [
          {
            "name": "page_no",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page No"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Page Size"
            }
          },
          {
            "name": "start_dt",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Dt"
            }
          },
          {
            "name": "end_dt",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Dt"
            }
          },
          {
            "name": "screen_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Screen Id"
            }
          },
          {
            "name": "manual_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Manual Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["admin-manual"],
        "summary": "Delete Manual",
        "description": "## 메뉴얼 다중 삭제\n\n### Args:\n    id_list: str = \"1, 2, 3\" (리스트 [1, 2, 3]가 편하실까요? 문자열 \"1, 2, 3\"이 편하실까요?)\n\n### Raises:\n    ManualException: id list's element is not number = id_list에 숫자가 아닌 게 끼어있을 때 (400, \"Bad Request\")\n\n### Returns:\n    status: (200, 'OK')\n    data: null",
        "operationId": "delete_manual_admin_manual__delete",
        "parameters": [
          {
            "name": "id_list",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id List"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/manual/embedding/{manual_id}": {
      "get": {
        "tags": ["admin-manual"],
        "summary": "Get One Manual",
        "description": "## 메뉴얼 id로 상세 조회 (좌측의 source/subject/content 페이지 별로 반환)\n## id 값은 페이지에 해당하는 id이며,\n## 상세 조회 시 그 메뉴얼에 해당하는 전체 페이지를 스크롤 형식으로 띄워줌.\n## 이 경우, 요청한 id가 처음으로 보여야 함\n\n### Args:\n    manual_id: int = 위 전체 조회 api response의 id 값 주시면 됩니다\n\n### Raises:\n\n### Returns:\n    status: (200, 'OK')\n    data: {\n        \"manual_name\": \"테스트1\",\n        \"page\": [\n            {\n                \"id\": 1,\n                \"source\": \"테스트1-1\",\n                \"subject\": \"어쩌구저쩌구\",\n                \"content\": \"어쩌구저쩌구\",\n                \"image\": null\n            },\n            {\n                \"id\": 2,\n                \"source\": \"테스트1-2\",\n                \"subject\": \"어쩌구저쩌구\",\n                \"content\": \"어쩌구저쩌구\",\n                \"image\": null\n            },\n            {\n                \"id\": 3,\n                \"source\": \"테스트1-3\",\n                \"subject\": \"어쩌구저쩌구\",\n                \"content\": \"어쩌구저쩌구\",\n                \"image\": null\n            }\n        ]\n    }",
        "operationId": "get_one_manual_admin_manual_embedding__manual_id__get",
        "parameters": [
          {
            "name": "manual_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Manual Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["admin-manual"],
        "summary": "Update Manual",
        "description": "## 메뉴얼 개별 수정 (메뉴얼 등록 팝업 2)\n## POST /embedding에서는 저장 버튼 눌러야 실제 테이블에 저장 되지만, 여기는 개별 수정 시 바로 반영된다\n\n### Args:\n    manual_id: 변경하려는 메뉴얼 번호 (path 말고 body 안에 넣는 게 나을까요?)\n    update_type: int = 변경 종류 (source = 0, subject = 1, content=2)\n\n### Raises:\n\n### Returns:\n    status: (200, 'OK')\n    data: null",
        "operationId": "update_manual_admin_manual_embedding__manual_id__patch",
        "parameters": [
          {
            "name": "manual_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Manual Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateManualRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/manual/preprocess/{set_id}": {
      "get": {
        "tags": ["admin-manual"],
        "summary": "Get One Preprocess",
        "description": "## /admin/manual/preprocess 요청 후 메뉴얼 등록 팝업 1의 값\n\n### Args:\n    set_id: int = POST /admin/manual/preprocess api response 값 주시면 됩니다\n\n### Raises:\n\n### Returns:\n    status: (200, 'OK')\n    \"data\": {\n        \"manual_name\": \"테스트1\",\n        \"page\": [\n            {\n                \"id\": 1,\n                \"source\": \"테스트1-1\",\n                \"subject\": \"어쩌구저쩌구\",\n                \"content\": \"어쩌구저쩌구\",\n                \"image\": null\n            },\n            {\n                \"id\": 2,\n                \"source\": \"테스트1-2\",\n                \"subject\": \"어쩌구저쩌구\",\n                \"content\": \"어쩌구저쩌구\",\n                \"image\": null\n            },\n            {\n                \"id\": 3,\n                \"source\": \"테스트1-3\",\n                \"subject\": \"어쩌구저쩌구\",\n                \"content\": \"어쩌구저쩌구\",\n                \"image\": null\n            }\n        ]\n    }",
        "operationId": "get_one_preprocess_admin_manual_preprocess__set_id__get",
        "parameters": [
          {
            "name": "set_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Set Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/manual/preprocess": {
      "post": {
        "tags": ["admin-manual"],
        "summary": "Create Preprocessing",
        "description": "## 메뉴얼 전처리 (메뉴얼 등록 1)\n\n### Args:\n    manual_name: 위의 POST /admin/manual/file의 response 값 넣어주시면 됩니다\n    pattern1, pattern1_1, pattern2: Optional[str] = {\"패턴\": \"입력 페이지\"}\n    현재 패턴은 1, 1_1, 2의 3종류이며, 제외 페이지는 exclude로 보내주실 수 있을까요?)\n    request 예시\n    {\n        \"manual_name\": \"~~\",\n        \"pattern1\": \"12, 20~23\",\n        \"pattern1_1\": \"1~11\",\n        \"pattern2\": \"13~14, 16~19\",\n        \"exclude\": \"24~40\"\n    }\n\n### Raises:\n\n### Returns:\n    status: (201, 'Created')\n    data: 생성된 메뉴얼 id 반환",
        "operationId": "create_preprocessing_admin_manual_preprocess_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePreprocessRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/manual/embedding": {
      "put": {
        "tags": ["admin-manual"],
        "summary": "Add Data",
        "description": "## 메뉴얼 수정 시 데이터 추가 (메뉴얼 등록 팝업 2)\n\n### Args:\n    manual_id: int (위에서 GET /admin/manual/embedding/{manual_id}의 manual_id 값)\n    source: Optional[str] = None\n    subject: Optional[str] = None\n    content: str\n    image_path: Optional[str] = None\n\n### Raises:\n\n### Returns:\n    status: (200, 'OK')\n    data: null",
        "operationId": "add_data_admin_manual_embedding_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddDataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["admin-manual"],
        "summary": "Create Collection",
        "description": "## 메뉴얼 전처리 완료 후 개별 수정 (메뉴얼 등록 2 팝업)\n\n### Args:\n    preprocess_id: int (GET /admin/manual/preprocess/{preprocess_id})의 return 값\n    updated_pages: List[dict] 변경할 값에 해당하는 데이터만 보내주시면 됩니다\n        \"updated_pages\": [\n            {\n                \"id\": 1,\n                \"source\": \"string\",\n                \"subject\": \"string\",\n                \"content\": \"string\",\n                \"image\": \"string\"\n            },\n            ...\n        ]\n\n### Raises:\n\n### Returns:\n    status: (201, 'Created')\n    data: null",
        "operationId": "create_collection_admin_manual_embedding_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCollectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["admin-manual"],
        "summary": "Delete Collection",
        "description": "## 벡터db 데이터 삭제\n## 데이터 임베딩용, 안 씀",
        "operationId": "delete_collection_admin_manual_embedding_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "deprecated": true
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddDataRequest": {
        "properties": {
          "manual_id": {
            "type": "integer",
            "title": "Manual Id"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "image_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Path"
          }
        },
        "type": "object",
        "required": ["manual_id", "content"],
        "title": "AddDataRequest",
        "example": {
          "content": "string",
          "image_path": "string",
          "manual_id": 1,
          "source": "string",
          "subject": "string"
        }
      },
...
      "CreateCollectionRequest": {
        "properties": {
          "set_id": {
            "type": "integer",
            "title": "Set Id"
          },
          "updated_pages": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ManualPage"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated Pages"
          }
        },
        "type": "object",
        "required": ["set_id"],
        "title": "CreateCollectionRequest",
        "example": {
          "set_id": 1,
          "updated_pages": [
            {
              "content": "string",
              "id": 1,
              "image_path": "string",
              "source": "string",
              "subject": "string"
            },
            {
              "content": "string",
              "id": 3,
              "image_path": "string",
              "source": "string",
              "subject": "string"
            }
          ]
        }
      },
      "CreatePreprocessRequest": {
        "properties": {
          "manual_name": {
            "type": "string",
            "title": "Manual Name"
          },
          "pattern1": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pattern1"
          },
          "pattern1_1": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pattern1 1"
          },
          "pattern2": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pattern2"
          },
          "exclude": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exclude"
          }
        },
        "type": "object",
        "required": ["manual_name"],
        "title": "CreatePreprocessRequest",
        "example": {
          "exclude": "",
          "manual_name": "00.기준정보_PHIS_설치및로그인.pdf",
          "pattern1": "",
          "pattern1_1": "",
          "pattern2": ""
        }
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ManualPage": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "image_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Path"
          }
        },
        "type": "object",
        "required": ["id", "source", "subject", "content", "image_path"],
        "title": "ManualPage"
      },
...
      "UpdateManualRequest": {
        "properties": {
          "update_type": {
            "type": "integer",
            "title": "Update Type"
          },
          "data": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "required": ["data"],
        "title": "UpdateManualRequest",
        "example": {
          "data": "string",
          "update_type": 1
        }
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}